add new binaryzation
