+++ include/decoder/absdecoder.hpp
@@ -45,8 +45,6 @@ namespace cv {
 
     void fillCounter(const std::vector<uchar> &row, int start, std::vector<int> &counters);
 
-    void adaptBinaryzation(InputArray src, OutputArray &dst);
-
     void cutImage(InputArray _src, OutputArray &_dst, RotatedRect rect);
 }
