+++ plot/data.csv
@@ -1,1 +1,1 @@
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-52,-253,-404,-407,-241,-45,-14,-10,-4,0,0,4,6,6,6,-1,-5,-1,-1,-3,-3,-3,-1,1,0,-2,-2,-2,-7,-16,-16,-7,8,16,15,11,3,3,4,-1,-5,-2,7,9,9,11,11,11,10,9,7,7,10,12,10,3,-8,-14,-11,-15,-18,-15,-9,-2,-1,-1,0,1,1,-1,-5,-5,0,3,5,7,5,4,4,-3,-12,-13,-10,-2,5,6,1,-3,-3,1,10,12,12,8,2,3,-3,-12,-14,-11,0,8,8,11,14,14,11,3,-2,-2,-6,-10,-12,-12,-5,-4,-8,-4,0,0,4,8,6,5,3,0,0,-5,-11,-11,-5,3,9,9,15,24,24,14,-1,-5,-4,0,4,4,6,8,8,5,1,0,0,1,2,3,1,-1,0,2,7,10,10,8,6,4,-3,-15,-23,-25,-9,8,8,6,2,2,4,2,0,0,2,4,2,2,2,0,0,-4,-8,-8,-7,-10,-11,-11,0,14,12,11,11,9,8,6,3,2,-3,-7,-7,-7,-3,0,0,-2,-4,-5,-5,-2,-2,-5,-7,-7,-6,2,14,16,16,15,13,14,11,8,8,4,1,1,0,1,2,2,6,10,7,8,13,13,12,9,6,6,0,-12,-18,-18,-9,0,0,2,2,2,4,5,9,9,7,8,8,5,1,0,0,-4,-8,-9,-6,-2,-2,-2,2,7,7,0,-7,-7,-2,11,23,23,10,0,0,3,12,16,12,-7,-25,-24,-8,12,17,17,15,12,14,6,-3,-3,0,9,14,14,13,12,14,8,0,-2,-4,-4,-7,-7,0,4,4,5,5,4,4,4,4,4,4,5,6,5,0,-4,-4,-4,-6,-8,-13,-26,-33,-35,-32,-25,-28,-8,32,46,47,47,43,41,29,10,3,4,6,9,10,4,-2,-2,-2,-3,-3,-3,2,8,7,8,10,13,12,11,13,13,7,1,3,2,0,1,1,16,33,33,17,-3,-12,-15,-13,-10,-10,-6,2,4,2,-3,-7,-8,-9,-9,-9,-3,11,17,15,18,18,15,9,1,-2,-4,-1,2,2,8,12,12,15,12,10,10,6,4,4,-1,-6,-9,-9,0,4,4,7,10,12,10,4,1,1,8,17,18,12,3,-1,-2,-3,-2,-2,2,12,12,8,4,0,0,0,-1,-1,-1,3,7,7,8,7,6,1,-2,-4,-7,-4,-1,-1,-1,4,6,2,3,4,4,4,0,0,1,-1,-2,-1,0,1,1,0,0,0,0,3,4,4,3,-1,-1,0,-4,-4,-4,-3,4,4,4,7,5,4,3,-3,-3,-4,-7,-7,-5,1,3,3,2,4,4,0,0,2,2,0,-2,-2,0,-2,-2,0,0,0,0,0,-1,-1,-1,-2,-5,-5,-3,-4,-4,-3,0,1,-1,1,1,0,1,-1,-1,2,7,8,7,5,2,4,-6,-19,-21,-20,-8,0,0,-8,-18,-21,-20,-13,-14,-14,-15,-21,-21,-35,-57,-66,-59,-20,9,9,8,8,10,9,7,8,9,-15,-38,-38,-25,2,15,13,2,-7,-7,-7,-6,-3,-3,-3,0,0,3,6,6,3,0,0,0,-5,-8,-8,-9,-4,-4,-8,3,12,12,5,-7,-9,-5,-4,-6,-6,-6,-6,-6,-3,0,0,0,2,4,4,-1,-6,-3,-3,-3,0,0,-3,-6,-6,-3,2,4,4,0,-4,-4,-6,-8,-7,-7,-9,-9,-8,-4,0,0,1,4,6,6,7,8,10,6,1,1,0,-10,-18,-15,-11,-8,-7,-4,5,12,12,6,0,0,-4,-11,-14,-14,-9,-6,-6,1,6,6,6,7,5,3,-6,-18,-18,-16,-13,-9,-8,1,9,10,6,-1,-2,-2,-8,-11,-8,-4,0,0,1,3,4,4,6,8,12,7,-2,-2,-2,-1,0,0,-3,-6,-6,-6,-5,-3,-2,0,2,2,1,-1,-2,-2,-5,-11,-11,-7,-6,-3,-3,-2,2,2,3,4,4,3,2,2,2,1,0,0,0,0,0,0,0,0,0,-1,-2,-2,-2,-1,0,0,4,8,11,8,2,2,2,1,0,0,-2,-4,-3,-5,-12,-14,-12,-9,-6,-6,-5,-2,0,0,-7,-14,-12,-1,12,14,11,3,-4,-4,0,3,0,4,12,16,16,6,0,1,3,5,4,5,12,21,20,0,-17,-19,-16,-4,4,4,13,20,21,19,12,11,13,26,35,34,24,9,4,-2,-20,-34,-36,-29,-20,-20,-17,-8,-2,-2,-4,-7,-8,-2,6,8,10,14,17,18,82,157,168,143,88,59,58,27,0,0,-2,-2,-5,-6,-6,-7,-7,-7,-6,-4,-12,-26,-34,-37,-85,-135,-135,-112,-72,-54,-50,-21,4,4,-4,-13,-14,-14,-20,-23,-23,-10,7,7,4,0,-2,-2,-2,-2,-2,1,3,4,9,13,15,17,15,12,14,13,6,2,2,2,3,2,-71,-191,-237,-210,-93,0,
+0,0,0,0,0,2,3,1,0,4,2,2,10,12,6,0,0,-3,-9,-14,-4,6,2,-1,-5,-5,-1,-1,-2,-11,-10,0,0,0,0,1,2,1,0,10,10,0,1,3,2,10,16,12,15,13,5,1,0,0,0,0,0,0,0,0,0,-12,-24,-12,0,0,0,0,1,5,9,1,-10,-6,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2,6,4,0,0,4,12,8,0,0,2,6,4,0,4,11,18,16,6,1,0,0,0,-10,-14,0,14,10,0,0,0,0,0,-5,-13,-16,-9,-6,-10,-7,-2,-2,-10,-10,-2,0,0,0,0,6,12,7,5,10,8,2,2,12,12,2,2,7,5,0,0,0,0,0,0,-3,-5,1,-1,0,6,2,0,0,0,0,0,0,0,-5,-13,-16,-9,-6,-6,-1,0,1,6,6,9,14,8,11,11,2,4,12,8,5,7,4,6,4,0,-6,-13,-7,0,0,0,0,-2,-10,-14,-11,-13,-19,-16,-6,-13,-27,-24,-13,-6,-6,-11,-8,-5,-4,0,-1,-2,-1,-3,-8,-11,-16,-31,-73,-127,-144,-102,-40,0,31,56,70,53,30,23,13,16,31,35,14,-9,-17,-21,-23,-19,-17,-12,-7,-7,2,7,-5,-29,-59,-59,-31,-1,45,42,7,-2,-15,-21,-56,-97,-84,-30,2,9,21,62,154,105,50,58,-11,-39,-36,-20,0,8,23,29,29,42,49,52,33,4,-8,-17,-18,-30,-34,-20,-17,-23,-21,-29,-39,-28,-13,2,48,104,140,209,392,297,23,-53,-115,-174,-104,-21,-26,36,94,58,13,5,-9,-36,-60,-63,-48,-14,7,6,-15,-48,-65,-83,-70,-50,-43,0,37,5,-105,-174,-85,69,103,52,20,-8,-10,-1,7,21,33,63,224,308,179,60,7,0,0,0,1,1,0,0,0,-1,-1,0,0,0,-1,-1,0,0,0,0,0,0,0,-1,-4,-3,-1,-1,0,1,1,0,0,-1,-1,0,0,0,0,0,-3,-4,-1,0,0,0,-4,-4,0,0,0,0,0,-1,-1,0,-2,-2,0,0,-1,-1,0,-1,-1,-3,-3,0,-1,-1,1,1,0,0,-1,-3,-5,-3,-2,-4,-6,-4,0,-1,-4,-4,-4,-3,-3,-3,-1,-4,-3,-1,-4,-5,-4,-3,-3,-4,-3,-2,-1,-2,-3,-6,-5,0,0,0,-3,-6,-3,0,0,0,0,0,-3,-7,-8,-7,-4,-2,-1,0,-1,-1,0,0,0,2,5,5,2,1,1,3,4,1,0,2,1,-4,-5,-2,-1,-2,-4,-5,-2,0,2,4,3,-6,-7,0,-1,-1,3,1,-2,3,7,8,3,-3,-4,-3,-1,-4,-7,-3,0,4,8,3,-8,-10,-7,-1,6,4,0,1,8,12,7,2,-1,-1,5,4,-1,-3,-2,1,0,-7,-8,-6,-1,8,9,3,-3,-4,-3,1,0,-3,-4,-2,4,5,2,-7,-13,-6,-2,-5,-11,-30,-57,-46,-20,-12,-3,1,17,46,52,27,-7,-20,7,23,12,12,10,2,4,10,8,6,5,1,0,0,0,0,3,3,0,0,0,0,0,-3,-4,-1,0,-4,-8,-4,0,3,6,3,-2,-4,-5,-8,-18,-28,-182,-343,-325,-215,146,451,348,131,26,4,5,11,9,1,-12,-247,-486,-434,-274,-92,12,29,33,56,112,187,225,237,268,251,123,20,5,5,5,1,1,2,2,5,4,-2,-4,-2,0,-7,-97,-337,-624,-469,-76,238,524,529,247,35,21,7,3,1,-4,-3,0,0,1,-2,-3,-2,-5,-5,-5,-14,-117,-372,-597,-393,-65,144,500,510,237,113,42,13,4,-6,-219,-522,-551,-280,-47,37,156,269,263,137,101,134,80,20,10,18,57,124,126,66,32,13,5,8,1,-10,-129,-433,-586,-319,19,215,391,451,272,83,34,23,13,0,4,4,0,0,0,0,0,0,0,0,3,3,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,-1,-1,0,0,0,0,0,0,0,0,0,0,-4,-4,0,0,-3,-11,-15,-17,-67,-205,-315,-270,-134,-93,-90,-39,-14,7,18,10,-5,-10,-9,-8,-5,-6,-4,-1,2,3,1,-4,-6,-2,-4,-8,-4,-10,-15,-8,-8,-8,-9,-5,3,1,2,33,50,41,39,25,22,22,14,9,2,-5,-2,2,-8,-9,-8,-11,-14,-19,-23,-24,-26,-26,-26,-34,-35,-28,-20,-15,-8,1,-3,-18,-41,-68,-72,-40,-11,5,36,75,58,36,46,28,-5,-51,-66,-33,11,24,54,75,37,19,21,21,17,9,8,17,14,13,11,-4,7,15,7,9,7,8,7,2,1,-1,-4,-18,-37,-69,-122,-140,-75,0,20,14,6,3,-2,-5,8,15,7,1,-5,-3,16,22,13,4,-1,7,6,2,0,-5,-1,2,-1,-4,-3,-2,-15,-22,-31,-57,-30,19,25,15,8,4,4,11,12,11,12,10,6,3,3,0,
